implement StdLibSpec for MongoDB JS evaluation
Factor out common std lib test fixture for MongoDB.
