+++ it/src/test/scala/quasar/physical/mongodb/MongoDbStdLibSpec.scala
@@ -1,0 +1,135 @@
+/*
+ * Copyright 2014–2016 SlamData Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package quasar.physical.mongodb
+
+import quasar.Predef._
+import quasar._, Planner.PlannerError
+import quasar.std._
+import quasar.fp._
+import quasar.fs.DataCursor
+import quasar.std.{StdLibSpec, StdLibTestRunner}
+import quasar.physical.mongodb.fs._, bsoncursor._
+import quasar.physical.mongodb.workflow._
+import WorkflowExecutor.WorkflowCursor
+
+import matryoshka._, Recursive.ops._
+import org.specs2.execute._
+import org.specs2.main.ArgProperty
+import scalaz._, Scalaz._
+import scalaz.concurrent.Task
+import shapeless.{Nat}
+
+/** Test the implementation of the standard library for one of MongoDb's
+  * evaluators.
+  */
+abstract class MongoDbStdLibSpec extends StdLibSpec {
+  args.report(showtimes = ArgProperty(true))
+
+  def shortCircuit[N <: Nat](backend: BackendName, func: GenericFunc[N]): Result \/ Unit
+
+  def compile(queryModel: MongoQueryModel, coll: Collection, lp: Fix[LogicalPlan])
+      : PlannerError \/ (Crystallized[WorkflowF], BsonField.Name)
+
+  def is2_6(backend: BackendName): Boolean = backend == TestConfig.MONGO_2_6
+
+  MongoDbSpec.clientShould { (backend, prefix, setupClient, testClient) =>
+    import MongoDbIO._
+
+    /** Intercept and transform expected values into the form that's actually
+      * produced by the MongoDB backend, in cases where MongoDB cannot represent
+      * the type natively. */
+    def massage(expected: Data): Data = expected match {
+      case Data.Time(time) => Data.Str(time.toString)
+      case _               => expected
+    }
+
+    /** Identify constructs that are expected not to be implemented. */
+    def shortCircuitLP: AlgebraM[Result \/ ?, LogicalPlan, Unit] = {
+      case LogicalPlan.InvokeF(func, _) => shortCircuit(backend, func)
+      case _ => ().right
+    }
+
+    def evaluate(wf: Crystallized[WorkflowF], tmp: Collection): MongoDbIO[List[Data]] =
+      for {
+        exc <- WorkflowExecutor.mongoDb.run.unattempt
+        v   <- exc.evaluate(wf, None).run.run(tmp.collection).eval(0).unattempt
+        rez <- v.fold(
+                _.map(BsonCodec.toData(_)).point[MongoDbIO],
+                c => DataCursor[MongoDbIO, WorkflowCursor[BsonCursor]].process(c).runLog.map(_.toList))
+      } yield rez
+
+    // TODO: this should probably have access to the arguments so it can inspect them as well
+    def check(arity: Int, prg: List[Fix[LogicalPlan]] => Fix[LogicalPlan]): Option[Result] =
+      prg((0 until arity).toList.map(idx => LogicalPlan.Free(Symbol("arg" + idx))))
+        .cataM[Result \/ ?, Unit](shortCircuitLP).swap.toOption
+
+    def run(args: List[Data], prg: List[Fix[LogicalPlan]] => Fix[LogicalPlan], expected: Data): Result =
+      check(args.length, prg).getOrElse(
+        (for {
+          coll <- MongoDbSpec.tempColl(prefix)
+          argsBson <- args.zipWithIndex.traverse { case (arg, idx) =>
+                      BsonCodec.fromData(arg).point[Task].unattempt.strengthL("arg" + idx) }
+          _     <- insert(
+                    coll,
+                    List(Bson.Doc(argsBson.toListMap)).map(_.repr)).run(setupClient)
+
+          qm  <- serverVersion.map(MongoQueryModel(_)).run(testClient)
+
+          lp = prg(
+                (0 until args.length).toList.map(idx =>
+                  Fix(StructuralLib.ObjectProject(
+                    LogicalPlan.Read(coll.asFile),
+                    LogicalPlan.Constant(Data.Str("arg" + idx))))))
+          t  <- compile(qm, coll, lp).point[Task].unattempt
+          (wf, resultField) = t
+
+          rez <- evaluate(wf, coll).run(testClient)
+
+          _     <- dropCollection(coll).run(setupClient)
+        } yield {
+          rez must_= List(Data.Obj(ListMap(resultField.asText -> massage(expected))))
+        }).unsafePerformSync.toResult)
+
+
+    val runner = new StdLibTestRunner with MongoDbDomain {
+      def nullary(
+        prg: Fix[LogicalPlan],
+        expected: Data): Result =
+        run(Nil, { case Nil => prg; case _ => scala.sys.error("") }, expected)
+
+      def unary(
+        prg: Fix[LogicalPlan] => Fix[LogicalPlan],
+        arg: Data,
+        expected: Data): Result =
+        run(List(arg), { case List(arg) => prg(arg) }, expected)
+
+      def binary(
+        prg: (Fix[LogicalPlan], Fix[LogicalPlan]) => Fix[LogicalPlan],
+        arg1: Data, arg2: Data,
+        expected: Data): Result =
+        run(List(arg1, arg2), { case List(arg1, arg2) => prg(arg1, arg2) }, expected)
+
+      def ternary(
+        prg: (Fix[LogicalPlan], Fix[LogicalPlan], Fix[LogicalPlan]) => Fix[LogicalPlan],
+        arg1: Data, arg2: Data, arg3: Data,
+        expected: Data): Result =
+        run(List(arg1, arg2, arg3), { case List(arg1, arg2, arg3) => prg(arg1, arg2, arg3) }, expected)
+    }
+
+    backend.name should tests(runner)
+  }
+}
